refactor: :art: refatoração da função writeFile
adicionado IF para distinguir se o DB possui ou não ou ID procurado

+++ src/helpers/writeFile.js
@@ -8,10 +8,19 @@ const filePath = path.resolve(__dirname, '..', 'talker.json');
 
 const writeFile = async (newTalker) => {
   const talkers = await readFile();
+  const doTalkerExists = talkers.find((talker) => talker.id === newTalker.id);
 
-  const updatedTalkers = JSON.stringify([...talkers, newTalker]);
-  
-  await fs.writeFile(filePath, updatedTalkers);
+  if (!doTalkerExists) {
+    const updatedTalkers = JSON.stringify([...talkers, newTalker]);
+    
+    await fs.writeFile(filePath, updatedTalkers);
+  } else {
+    const idToBeModified = talkers.findIndex((talker) => talker.id === newTalker.id);
+    talkers[idToBeModified] = newTalker;
+    const updateSpecificTalker = JSON.stringify(talkers);
+
+    await fs.writeFile(filePath, updateSpecificTalker);
+  }
 };
 
 module.exports = writeFile;
